shuffle around the repo so it's a bit easier to find files
