+++ src/components/confetti.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { Button } from "@mantine/core";
 
 const Confetti: React.FC = () => {
