+++ tests/03_irishbingo.spec.ts
@@ -1,9 +1,5 @@
 import { test, expect } from "@playwright/test";
-import {
-  SuitNames,
-  RankNames,
-  getCardName,
-} from "../src/components/games/card";
+import { SuitNames, RankNames, getCardName } from "@/components/carddeck";
 
 test("No card drawn on load", async ({ page }) => {
   await page.goto("/games/irishbingo");
