+++ types/index.d.ts
@@ -2,6 +2,6 @@ export {};
 
 declare global {
   interface Window {
-    ConfettiPage: any; // Adjust the type as needed
+    ConfettiPage: any;
   }
 }
